refactor(GatewayService): consolidate WebClient usage and remove redundant methods
Refactored the GatewayService to use a single WebClient instance per service (Server and Backend) instead of creating a new WebClient for each request. Removed redundant methods like `serverUri` and `backendUri` and streamlined the request handling logic. This improves maintainability and reduces code duplication.

+++ src/main/java/mindustrytool/servermanager/controller/ServerApiController.java
@@ -51,9 +51,9 @@ public class ServerApiController {
     }
 
     @GetMapping("total-player")
-    public Mono<Integer> getTotalPlayer() {
+    public Mono<Long> getTotalPlayer() {
         return serverService.getTotalPlayers()//
-                .onErrorReturn(0);
+                .onErrorReturn(0L);
     }
 
     @PostMapping("chat")
